requirements.txt
fastapi
uvicorn
psycopg2-binary
prometheus-client==0.20.0